fix: Number type uses IEEE 754 double instead of decimal
Reference compiler: IRNumberLit stores double, not decimal.
Lexer parses number literals as double. C# emitter outputs
double type and literals. x86-64 EmitExpr handles IRNumberLit
via BitConverter.DoubleToInt64Bits. Added text-to-double-bits
builtin (Text -> Integer, returns IEEE 754 bit pattern).

Backlog: text-to-double-bits bare metal implementation uses
__text_to_int as placeholder — proper float parser needed later.

+++ src/Codex.Emit.CSharp/CSharpEmitter.Utilities.cs
@@ -12,7 +12,7 @@ public sealed partial class CSharpEmitter
         return type switch
         {
             IntegerType => "long",
-            NumberType => "decimal",
+            NumberType => "double",
             TextType => "string",
             BooleanType => "bool",
             CharType => "long",
